refactor: 공용 Dropdown으로 Filter컴포넌트 리팩토링
