+++ src/components/Filter/styles.module.scss
@@ -1,0 +1,63 @@
+@use "../../assets/scss/base/mixins";
+
+.button {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  min-width: 13rem;
+  height: 4.2rem;
+  padding: 0 2rem;
+  border-radius: 1.2rem;
+  border: 1px solid var(--color-secondary-200);
+
+  @include mixins.mobile {
+    overflow: hidden;
+    justify-content: center;
+    min-width: 0;
+    width: 4.2rem;
+    padding: 0;
+  }
+
+  .label {
+    @include mixins.mobile {
+      display: none;
+    }
+  }
+
+  .arrow {
+    @include mixins.mobile {
+      display: none;
+    }
+  }
+
+  .sort {
+    display: none;
+
+    @include mixins.mobile {
+      display: block;
+    }
+  }
+}
+
+.list {
+  z-index: 100;
+  position: absolute;
+  right: 0;
+  top: 5rem;
+  width: 13rem;
+  border-radius: 1.2rem;
+  border: 1px solid var(--color-secondary-200);
+  background: #fff;
+
+  li + li {
+    border-top: 1px solid var(--color-secondary-200);
+  }
+
+  .item {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    height: 4.2rem;
+  }
+}
