--- src/components/Select/styles.module.scss
@@ -1,69 +1,0 @@
-@use "../../assets/scss/base/mixins";
-
-.select {
-  gap: 1rem;
-  position: relative;
-  color: var(--color-secondary-800);
-
-  .button {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    min-width: 13rem;
-    height: 4.2rem;
-    padding: 0 2rem;
-    border-radius: 1.2rem;
-    border: 1px solid var(--color-secondary-200);
-
-    @include mixins.mobile {
-      overflow: hidden;
-      justify-content: center;
-      min-width: 0;
-      width: 4.2rem;
-      padding: 0;
-    }
-
-    .label {
-      @include mixins.mobile {
-        display: none;
-      }
-    }
-
-    .arrow {
-      @include mixins.mobile {
-        display: none;
-      }
-    }
-
-    .sort {
-      display: none;
-
-      @include mixins.mobile {
-        display: block;
-      }
-    }
-  }
-
-  .list {
-    z-index: 100;
-    position: absolute;
-    right: 0;
-    top: 5rem;
-    width: 13rem;
-    border-radius: 1.2rem;
-    border: 1px solid var(--color-secondary-200);
-    background: #fff;
-
-    li + li {
-      border-top: 1px solid var(--color-secondary-200);
-    }
-
-    .item {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      width: 100%;
-      height: 4.2rem;
-    }
-  }
-}
